UI: Add support for redundant streams in bandwidth test
This commit adds support for redundant streams in
window-basic-auto-config. It adds a checkbox to the autoconfig page
for deciding whether redundant streams should be considered when
performing bandwidth tests. It also modifies the
TestBandwidthThread() to support streaming to multiple servers
simultaneously.

Implementation details:
In the TestBandwidthThread() implementation, the
connected/stopped state which is used to keep track of whether a given
output has started/stopped is replaced with a vector of states
tracked per output/server url. Instead of checking if a single
output is started/stopped as done previously, we now check if all
outputs are started/stopped. The start/stop states for a single
output along with the output itself and its callbacks are wrapped
in an OutputWrapper struct. In bandwidth test calculations,
preference is given to the redundant streams servers only if no
frames were dropped during the bandwidth test (and the redundant
streams checkbox was selected).

Since redundant streams can cause longer connection time. Servers with
redundant are preferred by virtually removing 5 ms to the biggest
connection time of the outputs.

Co-authored-by: Obinna Okechukwu <okecho@google.com>