CI: Add Flatpak aarch64
Building through QEMU emulation is really slow. So to avoid timing out
the job (6h), the original job is split in two.

The first job build dependencies (4-5h), if the cache already exist the
build is skipped.

And the second job build CEF and OBS Studio by relying on the cache of the
first job (2-3h).

This separation is done by modifying the manifest in the first job to
build only dependencies with a fake "obs" module/command.

aarch64 builds are not enabled by default on the main workflow, it can
be enabled with a dedicated label on PRs.