frontend,cmake: Get default video path on Linux with GLib
Qt does not always rely on xdg-user-dirs behavior to get the video
directory.
So the GLib API that relies on this behavior is used.

Note that GLib is already used by Qt itself for its event loop on Linux,
so adding GLib as a dependency of the UI is not an issue.